+++ agentsociety/llm/__init__.py
@@ -1,7 +1,7 @@
 """LLM related modules"""
 
 from .embeddings import SentenceEmbedding, SimpleEmbedding, init_embedding
-from .llm import LLM, LLMConfig, LLMProviderType
+from .llm import LLM, LLMConfig, LLMProviderType, monitor_requests
 
 __all__ = [
     "LLM",
@@ -10,4 +10,5 @@ __all__ = [
     "init_embedding",
     "LLMConfig",
     "LLMProviderType",
+    "monitor_requests",
 ]