+++ agentsociety/webapi/models/config.py
@@ -70,6 +70,7 @@ class ApiLLMConfig(BaseModel):
         for config in self.config:
             RealLLMConfig.model_validate(config)
 
+
 class MapConfig(Base):
     """Map model"""