+++ scripts/pyproject2requirements.py
@@ -1,7 +1,7 @@
 import tomli
 
-with open('pyproject.toml', 'rb') as f:
+with open("pyproject.toml", "rb") as f:
     data = tomli.load(f)
-deps = data['project']['dependencies']
-with open('requirements.txt', 'w') as f:
-    f.write('\n'.join(deps))
+deps = data["project"]["dependencies"]
+with open("requirements.txt", "w") as f:
+    f.write("\n".join(deps))